+++ include/databento/live.hpp
@@ -1,5 +1,6 @@
 #pragma once
 
+#include <chrono>
 #include <string>
 
 #include "databento/enums.hpp"  // VersionUpgradePolicy
@@ -29,6 +30,10 @@ class LiveBuilder {
   LiveBuilder& SetUpgradePolicy(VersionUpgradePolicy upgrade_policy);
   // Sets the receiver of the logs to be used by the client.
   LiveBuilder& SetLogReceiver(ILogReceiver* log_receiver);
+  // Overrides the heartbeat interval.
+  LiveBuilder& SetHeartbeatInterval(std::chrono::seconds heartbeat_interval);
+  // Overrides the gateway and port. This is an advanced method.
+  LiveBuilder& SetAddress(std::string gateway, std::uint16_t port);
   // Attempts to construct an instance of a blocking live client or throws an
   // exception.
   LiveBlocking BuildBlocking();
@@ -40,9 +45,12 @@ class LiveBuilder {
   void Validate();
 
   ILogReceiver* log_receiver_{};
+  std::string gateway_{};
+  std::uint16_t port_{};
   std::string key_;
   std::string dataset_;
   bool send_ts_out_{false};
   VersionUpgradePolicy upgrade_policy_{VersionUpgradePolicy::Upgrade};
+  std::chrono::seconds heartbeat_interval_{};
 };
 }  // namespace databento